Struct relay_dynamic_config::FeatureSet

pub struct FeatureSet(pub BTreeSet<Feature>);

A set of Features.

Tuple Fields
0: BTreeSet<Feature>
Implementations

impl FeatureSet

pub fn is_empty(&self) -> bool

Returns true if the set of features is empty.

pub fn has(&self, feature: Feature) -> bool

Returns true if the given feature is in the set.