Struct relay_dynamic_config::LimitedProjectConfig

pub struct LimitedProjectConfig {

    pub allowed_domains: Vec<String>,
    pub trusted_relays: Vec<PublicKey>,
    pub pii_config: Option<PiiConfig>,
    pub filter_settings: FiltersConfig,
    pub datascrubbing_settings: DataScrubbingConfig,
    pub dynamic_sampling: Option<SamplingConfig>,
    pub session_metrics: SessionMetricsConfig,
    pub transaction_metrics: Option<ErrorBoundary<TransactionMetricsConfig>>,
    pub metric_extraction: ErrorBoundary<MetricExtractionConfig>,
    pub metric_conditional_tagging: Vec<TaggingRule>,
    pub span_attributes: BTreeSet<SpanAttribute>,
    pub measurements: Option<MeasurementsConfig>,
    pub breakdowns_v2: Option<BreakdownsConfig>,
    pub performance_score: Option<PerformanceScoreConfig>,
    pub features: FeatureSet,
    pub tx_name_rules: Vec<TransactionNameRule>,
    pub tx_name_ready: bool,
    pub span_description_rules: Option<Vec<SpanDescriptionRule>>,

}

Subset of ProjectConfig that is passed to external Relays.

For documentation of the fields, see ProjectConfig.

Fields
allowed_domains: Vec<String>trusted_relays: Vec<PublicKey>pii_config: Option<PiiConfig>filter_settings: FiltersConfigdatascrubbing_settings: DataScrubbingConfigdynamic_sampling: Option<SamplingConfig>session_metrics: SessionMetricsConfigtransaction_metrics: Option<ErrorBoundary<TransactionMetricsConfig>>metric_extraction: ErrorBoundary<MetricExtractionConfig>metric_conditional_tagging: Vec<TaggingRule>span_attributes: BTreeSet<SpanAttribute>measurements: Option<MeasurementsConfig>breakdowns_v2: Option<BreakdownsConfig>performance_score: Option<PerformanceScoreConfig>features: FeatureSettx_name_rules: Vec<TransactionNameRule>tx_name_ready: bool

Whether or not a project is ready to mark all URL transactions as “sanitized”.

span_description_rules: Option<Vec<SpanDescriptionRule>>

Span description renaming rules.

These are currently not used by Relay, and only here to be forwarded to old relays that might still need them.

Implementations

impl LimitedProjectConfig

pub fn serialize<__S>( __self: &ProjectConfig, __serializer: __S ) -> Result<__S::Ok, __S::Error>where __S: Serializer,