Struct relay_dynamic_config::ProjectConfig

pub struct ProjectConfig {

    pub allowed_domains: Vec<String>,
    pub trusted_relays: Vec<PublicKey>,
    pub pii_config: Option<PiiConfig>,
    pub grouping_config: Option<Value>,
    pub filter_settings: FiltersConfig,
    pub datascrubbing_settings: DataScrubbingConfig,
    pub event_retention: Option<u16>,
    pub quotas: Vec<Quota>,
    pub dynamic_sampling: Option<SamplingConfig>,
    pub measurements: Option<MeasurementsConfig>,
    pub breakdowns_v2: Option<BreakdownsConfig>,
    pub performance_score: Option<PerformanceScoreConfig>,
    pub session_metrics: SessionMetricsConfig,
    pub transaction_metrics: Option<ErrorBoundary<TransactionMetricsConfig>>,
    pub metric_extraction: ErrorBoundary<MetricExtractionConfig>,
    pub span_attributes: BTreeSet<SpanAttribute>,
    pub metric_conditional_tagging: Vec<TaggingRule>,
    pub features: FeatureSet,
    pub tx_name_rules: Vec<TransactionNameRule>,
    pub tx_name_ready: bool,
    pub span_description_rules: Option<Vec<SpanDescriptionRule>>,

}

Dynamic, per-DSN configuration passed down from Sentry.

Fields
allowed_domains: Vec<String>

URLs that are permitted for cross original JavaScript requests.

trusted_relays: Vec<PublicKey>

List of relay public keys that are permitted to access this project.

pii_config: Option<PiiConfig>

Configuration for PII stripping.

grouping_config: Option<Value>

The grouping configuration.

filter_settings: FiltersConfig

Configuration for filter rules.

datascrubbing_settings: DataScrubbingConfig

Configuration for data scrubbers.

event_retention: Option<u16>

Maximum event retention for the organization.

quotas: Vec<Quota>

Usage quotas for this project.

dynamic_sampling: Option<SamplingConfig>

Configuration for sampling traces, if not present there will be no sampling.

measurements: Option<MeasurementsConfig>

Configuration for measurements. NOTE: do not access directly, use [relay_event_normalization::DynamicMeasurementsConfig].

breakdowns_v2: Option<BreakdownsConfig>

Configuration for operation breakdown. Will be emitted only if present.

performance_score: Option<PerformanceScoreConfig>

Configuration for performance score calculations. Will be emitted only if present.

session_metrics: SessionMetricsConfig

Configuration for extracting metrics from sessions.

transaction_metrics: Option<ErrorBoundary<TransactionMetricsConfig>>

Configuration for extracting metrics from transaction events.

metric_extraction: ErrorBoundary<MetricExtractionConfig>

Configuration for generic metrics extraction from all data categories.

span_attributes: BTreeSet<SpanAttribute>

The span attributes configuration.

metric_conditional_tagging: Vec<TaggingRule>

Rules for applying metrics tags depending on the event’s content.

features: FeatureSet

Exposable features enabled for this project.

tx_name_rules: Vec<TransactionNameRule>

Transaction renaming rules.

tx_name_ready: bool

Whether or not a project is ready to mark all URL transactions as “sanitized”.

span_description_rules: Option<Vec<SpanDescriptionRule>>

Span description renaming rules.

These are currently not used by Relay, and only here to be forwarded to old relays that might still need them.

Implementations

impl ProjectConfig

pub fn sanitize(&mut self)

Validates fields in this project config and removes values that are partially invalid.