Crate relay_dynamic_config

Protocol for dynamic configuration passed down to Relay from Sentry.

In contrast to static configuration files, parts of Relay’s configuration is generated by Sentry and polled by Relay in regular intervals. Dynamic configuration includes (but is not limited to)

rate limits and quotas,

feature flags,

settings that the user configured in Sentry’s UI.

Project Configuration

So far, the only scope of dynamic configuration is per [relay_auth::PublicKey] a.k.a. DSN. The schema for this configuration is defined in ProjectConfig.

Example Config

{
    "organizationId": 1,
    "config": {
        "excludeFields": [],
        "filterSettings": {},
        "scrubIpAddresses": False,
        "sensitiveFields": [],
        "scrubDefaults": True,
        "scrubData": True,
        "groupingConfig": {
            "id": "legacy:2019-03-12",
            "enhancements": "eJybzDhxY05qemJypZWRgaGlroGxrqHRBABbEwcC",
        },
        "blacklistedIps": ["127.43.33.22"],
        "trustedRelays": [],
    },
}

Structs

CustomMeasurementConfig

Configuration for extracting custom measurements from transaction payloads.

LimitedProjectConfig

Subset of ProjectConfig that is passed to external Relays.

ProjectConfig

Dynamic, per-DSN configuration passed down from Sentry.

SessionMetricsConfig

Configuration for metric extraction from sessions.

TaggingRule

Rule defining when a target tag should be set on a metric.

TransactionMetricsConfig

Configuration for extracting metrics from transaction payloads.

Enums

AcceptTransactionNames

Defines whether URL transactions should be considered low cardinality.

ErrorBoundary

Wraps a serialization / deserialization result to prevent error from bubbling up.

Feature

Features exposed by project config.

Functions

validate_json

Validate that the given JSON resolves to a valid instance of type S.